Move audio assets to own folder
