+++ src/ts/frontend/assets/audio/index.ts
@@ -15,7 +15,7 @@
 //  You should have received a copy of the GNU Affero General Public License
 //  along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-import { ILoadingProgressMonitor } from "./loadingProgressMonitor";
+import { ILoadingProgressMonitor } from "../loadingProgressMonitor";
 import { loadMusics, Musics } from "./musics";
 import { loadSounds, Sounds } from "./sounds";
 import { loadVoiceLines, SpeakerVoiceLines } from "./voiceLines";
